tools: refactor protoxform to distinct transform/pretty-print stages.
This patch moves protoxform to a protoc plugin that is solely in the
business of transforming FileDescriptorProtos to FileDescriptorProtos,
taking into account package status and performing vN -> v(N+1)
migrations.

A distinct tool, protoprint.py, which has the guts of the former
protoxform.py (basically no changes), is then invoked by proto_sync.py
during source tree copy-back.

The rationale for this is split is that proto_sync.py needs to be able
to combine both the v2 next major version candidate and the hand edited
v3 protos when generating the v3 shadow, once v3 becomes canonical. This
is easiest to do when we operate on FileDescriptorProtos.

Risk level: Low
Testing: proto_format.sh (no changes).
