+++ tools/protoxform/BUILD
@@ -8,11 +8,32 @@ py_binary(
         "protoxform.py",
         "utils.py",
     ],
-    data = ["//:.clang-format"],
     python_version = "PY3",
     visibility = ["//visibility:public"],
     deps = [
         "//tools/api_proto_plugin",
+        "//tools/type_whisperer:api_type_db_proto_py_proto",
+        "@com_envoyproxy_protoc_gen_validate//validate:validate_py",
+        "@com_github_cncf_udpa//udpa/annotations:pkg_py_proto",
+        "@com_google_googleapis//google/api:annotations_py_proto",
+        "@envoy_api_canonical//envoy/annotations:pkg_py_proto",
+    ],
+)
+
+py_binary(
+    name = "protoprint",
+    srcs = [
+        "options.py",
+        "protoprint.py",
+        "utils.py",
+    ],
+    data = [
+        "//:.clang-format",
+        "//tools/type_whisperer:api_type_db.pb_text",
+    ],
+    python_version = "PY3",
+    visibility = ["//visibility:public"],
+    deps = [
         "//tools/type_whisperer",
         "//tools/type_whisperer:api_type_db_proto_py_proto",
         "@com_envoyproxy_protoc_gen_validate//validate:validate_py",
