+++ tools/protoxform/protoxform_test_helper.py
@@ -4,8 +4,11 @@ from run_command import runCommand
 
 import logging
 import os
+import pathlib
 import re
+import subprocess
 import sys
+import tempfile
 
 
 def PathAndFilename(label):
@@ -44,11 +47,27 @@ def GoldenProtoFile(path, filename, version):
   return os.path.abspath(base)
 
 
-def ResultProtoFile(path, filename, version):
+def ProtoPrint(src, dst):
+  """Pretty-print FileDescriptorProto to a destination file.
+
+  Args:
+    src: source path for FileDescriptorProto.
+    dst: destination path for formatted proto.
+  """
+  print('ProtoPrint %s' % dst)
+  contents = subprocess.check_output([
+      'bazel-bin/tools/protoxform/protoprint', src,
+      './bazel-bin/tools/protoxform/protoprint.runfiles/envoy/tools/type_whisperer/api_type_db.pb_text'
+  ])
+  pathlib.Path(dst).write_bytes(contents)
+
+
+def ResultProtoFile(path, tmp, filename, version):
   """Retrieve result proto file path. In general, those are placed in bazel artifacts.
 
   Args:
     path: target proto path
+    tmp: temporary directory.
     filename: target proto filename
     version: api version to specify target result proto filename
 
@@ -59,7 +78,9 @@ def ResultProtoFile(path, filename, version):
   base += os.path.join(path, "protos")
   base += os.path.join(base, path)
   base += "/{0}.{1}.proto".format(filename, version)
-  return os.path.abspath(base)
+  dst = os.path.join(tmp, filename)
+  ProtoPrint(os.path.abspath(base), dst)
+  return dst
 
 
 def Diff(result_file, golden_file):
@@ -91,15 +112,16 @@ def Run(path, filename, version):
     result message extracted from diff command
   """
   message = ""
-  golden_path = GoldenProtoFile(path, filename, version)
-  test_path = ResultProtoFile(path, filename, version)
+  with tempfile.TemporaryDirectory() as tmp:
+    golden_path = GoldenProtoFile(path, filename, version)
+    test_path = ResultProtoFile(path, tmp, filename, version)
 
-  status, stdout, stderr = Diff(test_path, golden_path)
+    status, stdout, stderr = Diff(golden_path, test_path)
 
-  if status != 0:
-    message = '\n'.join([str(line) for line in stdout + stderr])
+    if status != 0:
+      message = '\n'.join([str(line) for line in stdout + stderr])
 
-  return message
+    return message
 
 
 if __name__ == "__main__":
