+++ tools/testdata/protoxform/envoy/v2/discovery_service.proto.next_major_version_candidate.envoy_internal.gold
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package envoy.v3;
 
-import "";
+import "envoy/api/v2/discovery.proto";
 
 import "google/api/annotations.proto";
